lib64/os: Optimize thread creation.
Thread creation resulted in a lot of context switches, as
the current thread was always flushed out to RAM, the new
thread's state was written, and one of the two threads was
loaded (depending upon the priority).

Now, if the new thread has higher priority, the current
thread is saved and the new thread in setup in-place. In
addition, if the new thread has a lower or equal priority,
then no context switching at all occurs - the thread is
simply queued for later.

Signed-off-by: Tyler J. Stachecki <stachecki.tyler@gmail.com>