libn64/os: Fix cache bugs.
In syscall.s, the cache instruction offset was harmless,
but the offset being wrong resulted in doing nothing of
value.

In context.s, we were marking a cache line dirty that
should not have been. This sometimes resulted in the value
of the thread's MI_INTR_REG getting overwritten with
garbage, at which point unexpected interrupts would start
firing and the framework would freeze.

Signed-off-by: Tyler J. Stachecki <stachecki.tyler@gmail.com>