Faciliate loading of external assets.
Maintain a 'filesystem' directory placed within the
the source directory - whenever make is run, files in
this directory are assembled into the final image.

An external header file is generated with offsets and
sizes of each of the files packaged. This allows for
compile-time resolution of addresses within the cart
(as opposed to doing costly dynamic lookups based on
directory and filename).

Signed-off-by: Tyler J. Stachecki <stachecki.tyler@gmail.com>