Filesystem creation tool and message queues. (#2)
* Faciliate loading of external assets.

Maintain a 'filesystem' directory placed within the
the source directory - whenever make is run, files in
this directory are assembled into the final image.

An external header file is generated with offsets and
sizes of each of the files packaged. This allows for
compile-time resolution of addresses within the cart
(as opposed to doing costly dynamic lookups based on
directory and filename).

Signed-off-by: Tyler J. Stachecki <stachecki.tyler@gmail.com>

* Add message allocator interfaces.

The message allocators were already in-use within the
kernel - these syscalls just crack them open for direct
allocations.

The message allocators provide a means for quickly
obtaining 16-byte structures that are aligned on data
cache line boundaries.

Signed-off-by: Tyler J. Stachecki <stachecki.tyler@gmail.com>

* Add support for "unbound" MP-SC message queues.

Add support for message queues which are not directly tied
to a thread and can be passed around as needed. Any thread
can send a message on a message queue, but only the thread
which created the message queue can recv on it.

Unlike thread-specific message queues, the receiver can opt
to block on the queue (suspend himself if no message are
available) or block until a message becomes available on
the queue.

Signed-off-by: Tyler J. Stachecki <stachecki.tyler@gmail.com>