macOS Build Support (#8)
* Removed unnecessary static linking. Now builds on macOS.

* Updated build instructions for Mac and Linux

For Linux, I removed yacc from the list of packages to install with apt. As far as I can tell, yacc is provided as an alias by bison and the package 'yacc' is not found in the repositories.

For Mac, I added instructions to install the prerequisites via the homebrew package manager.

* Updated build prerequisites

Added flex as a prerequisite.